Build out `new` command
The `quire new` command is now working for basic usage. When this command runs,

* it downloads the starter repo defined at the top of the file
* it downloads the theme repo defined at the top of the file
* it installs NPM depencencies in each folder
* it prints some helpful messages to the user

The cmd_check file has been changed to cmd_missing and should be used like a
Ruby guard clause (if commandMissing is true, thrown an error and exit;
otherwise continue and use the command in question).

Because this is a basic script, I'm using synchronous methods to keep out of
callback hell. This shouldn't be a problem for the time being.