WIP: reorganize project into ES6 class
Trying to re-write things so that this code is easier to reason about. For now
I've settled on the "EventEmitter" pattern: creating a custom ES6 class and
inheriting from Node's EventEmitter for the core CLI object. The various
commands live in different instance methods on the object, and messages are sent
to this object from the user via .emit() events. The CLI listens for a 'preview'
event and fires the preview() method when it hears one, etc.

"Private" methods are not really supported but by convention I'm prepending an
underscore onto those method names to imply that they are only meant for private
use. This is where most of the code from the "util" file will move to.