Re-write PDF generation methods
PDF generation now follows this approach:

1. Prince hits the live preview on localhost instead of the static built files.
This allows the internal links to work more smoothly and seems to be more
consistent in handling CSS.

2. The _getPathForChapter() method has been modified to return the urls as they
would be on the server.

3. The actual call to Prince CLI happens in a setTimeout callback. But ideally
this should be re-written as a promise chain. This would also allow the dev
server to auto-shutdown when the PDF is done.