Move range exclusions out of logic, take out `index.md` requirement
This partial ranges across all pages, but has a number of large
exclusions including `online: false`. Moves these exclusions out of
logic, lines 38-40.

This commit also rewrites some of the logic so that a section doesn’t
need to have an `index.md` file to list properly. When no `index.md`
file is present, the .Section name is included with no link for the
“full” version of the contents list, and with a link to the first page
in the section for the “short” version of the contents list.