Use Bulma's .content component for text content areas
Bulma's .content styles go a long way to doing what we need, so we will use them
and override only what is needed.

One such override is the ability for the user to define a value for
$quire-page-paragraph-style. By default paragraphs are all flush left and broken
up by whitespace. By setting this value to 'indent', an alternate style is used
where vertical space is removed and horizontal text-indent is added to
paragraphs which immediately follow other paragraph elements.