Update CLI to be version-aware
There is now a global config var that tells the CLI which version of the starter
kit to download. This should only be bumped when the version of the CLI tool
changes, so that any development changes on the starter kit do not break things
downstream for users.

The Git workflow has also been slightly improved. The CLI now
clones (recursively) the starter kit at the desired release tag. Since this
leaves the new repo in headless state by default, a new local master branch is
created as well.