Fix and expand citation and bibliography handling
The q-cite shortcode creates a pop-up citation with a little added css.
Additional citations can be added to the automatically generated
page-level bibliographies with `additional_references` in the page
yaml. Added some inline documentation to the big pieces.

Couple issues:
1. The pop-ups could probably be handled better, but I wanted to keep
it as minimal as possible for now.
2. There’s a chunk of code that is almost repeated verbatim in q-cite
and again in additional-references.html but I was having trouble
pulling the chunk out as it’s own partial.
3. In the jump links in the page header, the Bibliography one will only
be generated IF there’s a bibliography to link to, but we can’t do the
same with Notes. So if there are no footnotes, the Notes jump link is
still there but non-functional.