Accessibility improvement: Keyboard nav & menu state
This change improves accessibility by making the site menu behave better with
users who chose to navigate via keyboard (tabbing through links). By using
display:none when the menu is closed and display:block when it is open, the user
is only able to tab through the links inside the menu if they have explicitly
chosen to expand it. Otherwise they continue tabbing through the rest of the
page. This works regardless of the underlying source order.

This latter fact enables us to also fix a layout bug that was most easily solved
by the CSS adjacent sibling combinator (adjusting the width of the primary
region when the menu region was expanded, so that one is not hidden by the
other).