Update quire-contributors shortcode and styles
The quire-contributors shortcode is one of the most complex components of the
template currently, since it supports several different modes of display.

This commit does not change any of the underlying logic (though some of this
could probably be simplified in the future), but it does attempt to make the
code more legible by adding some comments and formatting improvements.

The only area that markup has been changed is in the "bio" section, to support
use of Bulma's media object. A few stylistic overrides have been added in the
quire-contributors.scss file as well.