Rename Quire class => CLI
Begin code reorganization. The first step is to split up the basic CLI from the
project-handling code into a CLI class and a Project class, respectively.
Creating a new boilerplate project from scratch is the purview of the former;
managing existing projects with arbitrary configuration falls under the domain
of the latter.