Make image lookup more bulletproof
Right now `imageDir` could be modified arbitrarily – or even deleted – by the
user, and Hugo wouldn't care. To try to handle all the possible cases here, this
is how I'm handling it.

1. At the top of the relevant template file, we need to check if the `imageDir`
has been set in the site config. We can then store a scratch value equal to the
value the user provided here, or an empty string if it doesn't exist:

```
{{ if isset $.Site.Params "imagedir" }}
{{ $.Scratch.Set "imageDir" $.Site.Params.imageDir }}
{{ else }}
{{ $.Scratch.Set "imagedir" "" }}
{{ end }}
```

(Note that `imagedir` is **lowercase** in the `isset` clause above; Hugo stores
all site-level configuration keys as lowercase)

2. Whenever we need to output the image location in a `src` attribute, we should
use Hugo's built-in `relURL` or `absURL` functions instead of building URLs
manually, as this is much less error-prone. Go templates do not have a simple
string concatenation function like Ruby or JS do, so we have to use
`printf` (https://gohugo.io/functions/printf/) to add strings together.

```
{{ with .Params.image }}
{{- $imgPath := printf "%s/%s" ($.Scratch.Get "imageDir") . -}}
<div class="quire-entry__image">
  <img src="{{ $imgPath | relURL }}" alt="" />
</div>
{{ end }}
```