Allow for background images, and fix fullheight issue
Background images can be added to the cover, page headers, and section-head "splash" page by adding an `image` to the page YAML for that page.

This also fixes an issue where Bulma's `is-fullheight` class was set to 100vh but did not take into account our navbar, and so fullheight elements extended beyond the height of the browser window.