Convert `content` page type to work in place of `section-head`
Previously, we had defined a `section-head` page type that was a langing page for sub-directory sections. However, as implemented, the landing page was basically just a table of contents for the section. This commit expands the functionality of the `content` page type so that it can work on sections, and can be formatted one of four ways by adding a `class` to the page YAML. The class can display the contents lists as a `brief` list, as a regular `list` (default), as a list with `abstract`s or as a visual `grid`. This also means that section landing pages needn't be limited to only `section-head` types, but any type of Quire page.