Add support for embedding website as figure
Uses an iframe. Can be listed in figures.yml as usual, with a `media_type` of "website", and a `media_id` equal to the URL. As with video figures, a static fallback image of the site can be included under `src`. When a fallback is not provided, an SVG icon placeholder is inserted for the pdf edition. The caption in the print edition will include a link when `websiteFigureFallbackText` is set to "true" in `config.yml`