Convert figure groups to flexbox grid
Users can still specify a `grid` for the `qifigure-group` shortocde, but now, when the number of images don't match the grid, the remaining images in a row fil the space. The one issue is that PrinceXML doesn't support flexbox, so there, the last images retain their grid widths and float left.