Regularize scss file structure and variable handling
Every .scss file now has a consistent structure of comments and headings and includes a list of every variable used in that file and its default value. The values of the default variables in the individual files, also match up exactly with the starting values in the variables.scss file, where users will be directed to make changes for customization.