Tidy q-contributor shortcode and add id support
This tidies up the q-contributor shortcode significantly (pulling part of it into a new partial) and also adds support to use `id` references in page YAML that can refer back to contributors listed in the publication.yml file. In this way, contributor info can be kept in one location, without duplicating information.