Remove use of "Params.epub true" and urlize Title IDs
There were two remaining uses of the `{{ if eq .Site.Params.epub true }}` logic. One of them could be fixed with the `remove-from-epub` class, and the other needed to have some markup change, placing an `id` on a span element rather than on a `dt` or `dd` element which Pandoc was stripping for some reason.

This also urlizes the use of the page `.Title` in some IDs to be consistent with the IDs on all other headings that come out of the markdown processor.