collapse contents-list and arrow-list back into one partial
The newly introduced arrow-list duplicated a lot of code that is in the original contents-list partial. It also introduced a `section-indent` class to create visually nested lists, rather than the semantically nested lists the contents-list partial created. This commit eliminates the arrow-list partial in favor of the original version. The necessary template and style changes were then made to the conents-list partial to get the desired output in all versions of the table of contents as well as the side menu.