Add site and page title partials
Titles and subtitles are concatenated throughout Quire's templates. This is mostly with a colon, but in cases where the title ends in an exclamation point or question mark, no colon is included. This requires an extra bit of logic and instead of repearing that a dozen times across the site, it's now available in two simple partials, one for the site title and another for the page title