Standardize footer buttons logic, avoid certain pages
This is parallel to the handling of the next prev links in the navbar. Instead of using range, this uses only .Next and .Prev but also checks to eliminate pages that are marked as `online: false` or are `type:data`. Also fixes an issue with the "image-control-icon" that wasn't generating consistently.