Initial commit Cite this Page feature
Handles basic Chicago and MLA formatting. However, it doesn't account for name combinations that might end with a period (Keaton, Alex P., Sammy Davis Jr.) and so in some cases may end up including two periods in a row.

Also will require a bit of javascript to populate the Access date.

And finally, the styling has been done ad hoc and in line for now until design can look it over and finalize.