Fix nav bar prev and next bug
The way the logic was working in the navbar previously, if the .Next or .Prev page (as defined by Hugo) included `online: false` no heading and link was included in the navbar. This new method looks at the entire list of online pages that is either above (next) or below (previous) the current page's .Weight, and then links to the first page in the list (next) or the last page in that list (previous).