Strip formatting from <title>
If you use markdown in the title of a page, it was also being used in the `<title>` tag and showing up raw in browser tabs, and so then search engines, etc. This strips that formatting.

This also will only include page titles on page other than the homepage. The homepage will just be the publication title.