Wrap text version as button rather than link and remove citationPopupLinkText
I think it's good to use <button> instead of <a> when it changes the state of something on the page (like it dow does with opening the tooltip) rather than when it takes you to a new location, as it used to.

After thinking about it more and after consulting with our editors, I'm also removing the link to the page bibliography that I'd asked for. It's too visually important, and has very little use case.