Structure initial version of new shortcode
The primary goal is to eliminate a lot of the repetition in the current shortcode, and to use partials to make the shortcode more modular and easier to add to.

This first pass shows how the figure attributes are defined either from the shortcode itself if a src is given, or from a figures.yml file.

It include basic figure and caption partials, with room for more for each media type.