Add figure label partial and rewrite styles to match
This partial will also be used for the other figure media_types. The original version of the labels had significantly different markup and class names for the two versions of the labels (appearing on top of the figure in the upper left corner, or appearing below the image to the left of the caption), this new partial uses the same markup for each version and the styles were rewritten accordingly.

Due to the way Pandoc processes files for the EPUB, the label had to be inserted into the figure caption partial for EPUB output in order to get it to not be deleted/ignored by Pandoc.