Update README.md
Still need to:

Update with Quire website when that is available.
Add deployment information (coming soon).
Redirect to correct Contributing Guidelines file
Add logo to repo

Let me know if you have any feedback, especially when it comes to the order / hierarchy of everything. Any additions / subtractions? The other part I'm unsure of is the handling of the Contributing Guidelines. Should this be further abbreviated?