Remove use of contentsType
Removes tocType and menuType as configuration parameters and instead relies on the class of the contents page itself (in place of tocType), and a new menuClass parameter (in place of menuType). And the accepted values for menuClass mirror those for the contents page class (brief and list) rather than using short and full as before.