Update to eleventy 1.0.0-beta.6
Refactors npm scripts to set an ELEVENTY_ENV variable and use npx
to run local binaries.

Adds a prebuild script to ensure that the output directory is clean.
Adds a serve script to run a web server for the local output directory.
Adds a start script to build and serve a local production build.