Remove `pageFigures` getter from `lightbox` lit-element, replacing with `figures` passed as a prop
`pageFigures` was a bit of a hack, as the way `lightbox` and `modal`
components were previously registered did not provide access to `page`
data, so instead they received ALL figures from `globalData`, and
filtered them based on the `id` attributes of query-selected `.q-figure`
elements on the page. Now they are only passed figures on the current
page.