Fix: Refactor markdownify filter to conditionally render block or inline
Refactors the universal template filter `markdownify` to conditionally use the markdown-it `renderInline` method when the markdown string contains new-line characters. This allows yaml values to contain markdown lists as multiline strings, for example an entry in the `objects.yaml` `object_list` with the following key will be rendered as an HTML unordered list by the `tombstone` component.